requirements.txt
requests
tenacity
twine
pre-commit
jinja2
certifi
pyyaml
toml
click
cryptography
setuptools
coverage
pykwalify
jsonschema
python-dotenv
dynaconf
urllib3[secure]
requests[security]
hvac
hatch
